refactor: Extract LogPanel from LogTab
