+++ src/ui/panel/mod.rs
@@ -1,3 +1,5 @@
 mod details_panel;
+mod log_panel;
 
 pub use details_panel::{DetailsPanel, DetailsPanelEvent};
+pub use log_panel::LogPanel;
